Changed files for vercel backend integration
